src/requirements.txt
streamlit==1.32.0
Pillow==10.2.0
requests==2.31.0
folium==0.16.0
streamlit-folium==0.18.0
exifread==3.0.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-telegram-bot==20.8
